# docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "title": "Shift Scheduler Service API",
        "contact": {
            "name": "Yunus Emre Alpu",
            "url": "https://yunusemrealpu.netlify.app",
            "email": "YunusAlpu@icloud.com"
        },
        "version": "1.0.0"
    },
    "basePath": "/shift-scheduler-service",
    "paths": {
...
        "/shift-schedules": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "get all shift schedules",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Shift"
                ],
                "summary": "get all shift schedules",
                "responses": {
                    "200": {
                        "description": "get all shift schedules successfully",
                        "schema": {
                            "$ref": "#/definitions/handlers.RespondJson"
                        }
                    },
                    "400": {
                        "description": "cannot get all shift schedules due to invalid request body",
                        "schema": {
                            "$ref": "#/definitions/handlers.RespondJson"
                        }
                    },
                    "422": {
                        "description": "cannot get all shift schedules due to invalid request body",
                        "schema": {
                            "$ref": "#/definitions/handlers.RespondJson"
                        }
                    },
                    "500": {
                        "description": "cannot get all shift schedules due to internal server error",
                        "schema": {
                            "$ref": "#/definitions/handlers.RespondJson"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "create a new shift schedule",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Shift"
                ],
                "summary": "create a new shift schedule",
                "parameters": [
                    {
                        "description": "Alias",
                        "name": "alias",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "Description",
                        "name": "description",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "Frequency",
                        "name": "frequency",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "description": "Start Date",
                        "name": "start_date",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "End Date",
                        "name": "end_date",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "Year",
                        "name": "year",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "description": "Status",
                        "name": "status",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "description": "Organization",
                        "name": "organization",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object"
                        }
                    },
                    {
                        "description": "Manager",
                        "name": "manager",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object"
                        }
                    },
                    {
                        "description": "Users",
                        "name": "users",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object"
                        }
                    },
                    {
                        "description": "Shifts",
                        "name": "shifts",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successfully created shift schedule",
                        "schema": {
                            "$ref": "#/definitions/handlers.RespondJson"
                        }
                    },
                    "400": {
                        "description": "cannot create shift schedule due to invalid request body",
                        "schema": {
                            "$ref": "#/definitions/handlers.RespondJson"
                        }
                    },
                    "422": {
                        "description": "cannot create shift schedule due to invalid request body",
                        "schema": {
                            "$ref": "#/definitions/handlers.RespondJson"
                        }
                    },
                    "500": {
                        "description": "cannot create shift schedule due to internal server error",
                        "schema": {
                            "$ref": "#/definitions/handlers.RespondJson"
                        }
                    }
                }
            }
        },
        "/shift-schedules/deleted": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "get all deleted shift schedules",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Shift"
                ],
                "summary": "get all deleted shift schedules",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Shift Schedule ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "get all shift schedules successfully",
                        "schema": {
                            "$ref": "#/definitions/handlers.RespondJson"
                        }
                    },
                    "400": {
                        "description": "cannot get only deleted shift schedules due to invalid request body",
                        "schema": {
                            "$ref": "#/definitions/handlers.RespondJson"
                        }
                    },
                    "422": {
                        "description": "cannot get only deleted shift schedules due to invalid request body",
                        "schema": {
                            "$ref": "#/definitions/handlers.RespondJson"
                        }
                    },
                    "500": {
                        "description": "cannot get only deleted shift schedules due to internal server error",
                        "schema": {
                            "$ref": "#/definitions/handlers.RespondJson"
                        }
                    }
                }
            }
        },
...
        "/shift-schedules/{id}": {
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "delete a shift schedule",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Shift"
                ],
                "summary": "delete a shift schedule",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Shift Schedule ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successfully deleted shift schedule",
                        "schema": {
                            "$ref": "#/definitions/handlers.RespondJson"
                        }
                    },
                    "400": {
                        "description": "cannot delete shift schedule due to invalid request body",
                        "schema": {
                            "$ref": "#/definitions/handlers.RespondJson"
                        }
                    },
                    "422": {
                        "description": "cannot delete shift schedule due to invalid request body",
                        "schema": {
                            "$ref": "#/definitions/handlers.RespondJson"
                        }
                    },
                    "500": {
                        "description": "cannot delete shift schedule due to internal server error",
                        "schema": {
                            "$ref": "#/definitions/handlers.RespondJson"
                        }
                    }
                }
            }
        },
...
        "/shift-schedules/{year}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "get a shift schedules by year",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Shift"
                ],
                "summary": "get a shift schedules by year",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Shift Schedule Year",
                        "name": "year",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "get shifts by year successfully",
                        "schema": {
                            "$ref": "#/definitions/handlers.RespondJson"
                        }
                    },
                    "400": {
                        "description": "cannot get shifts schedule by year due to invalid request body",
                        "schema": {
                            "$ref": "#/definitions/handlers.RespondJson"
                        }
                    },
                    "422": {
                        "description": "cannot get shifts schedule by year due to invalid request body",
                        "schema": {
                            "$ref": "#/definitions/handlers.RespondJson"
                        }
                    },
                    "500": {
                        "description": "cannot get shifts schedule by year due to internal server error",
                        "schema": {
                            "$ref": "#/definitions/handlers.RespondJson"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "handlers.RespondJson": {
            "type": "object",
            "properties": {
                "intent": {
                    "type": "string"
                },
                "message": {},
                "status": {
                    "type": "boolean"
                }
            }
        },
...
    }
}